11/05/21 18:18 - Correction Dijkstra

+++ be-graphes-algos/src/main/java/org/insa/graphs/algorithm/shortestpath/Label.java
@@ -9,22 +9,22 @@ public class Label implements Comparable<Label>{
 	
 	protected boolean marque ;
 	
-	protected float cout ;
+	protected double cout ;
 	
 	protected Arc pere ;
 	
     public Label(Node courant) {
     	this.sommetCourant = courant ;
     	this.marque = false ;
-    	this.cout = Float.MAX_VALUE ;
+    	this.cout = Double.MAX_VALUE ;
     	this.pere = null ;
     }
     
     public Node getSommet() {return this.sommetCourant ;}
     
-    public float getCost() {return this.cout ;}
+    public double getCost() {return this.cout ;}
     
-    public void setCost(float cout) {this.cout = cout ;}
+    public void setCost(double cout) {this.cout = cout ;}
     
     public boolean getMark() {return this.marque ;}
     
@@ -35,7 +35,7 @@ public class Label implements Comparable<Label>{
     public void setFather(Arc pere) {this.pere = pere;}
     
     public int compareTo(Label other) {
-        return Float.compare(getCost(), other.getCost());
+        return Double.compare(getCost(), other.getCost());
     }
 
 }
